postgres_to_es/requirements.txt
python-dotenv==0.20.0
psycopg2-binary==2.9.3
elasticsearch==7.7.0
redis==4.2.2
isort==5.10.1